backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
alembic
pydantic
email-validator
python-jose
passlib[bcrypt]
python-multipart
pillow
pillow-heif
httpx
openpyxl
pytest
